% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{artificial_networks}
\alias{artificial_networks}
\title{200 Simulated Networks of order 2000 with Polylogarithmic (0.1, 2)
Degree Distributions}
\format{a list containing 200 network objects. Each network object is a list
with three elements:
\describe{
   \item{edges}{edgelist of the network. A two column
     \code{matrix} where each row is an edge.}
   \item{degree}{degree sequence of the network, which is
     an \code{integer} vector of length \eqn{n}.}
   \item{n}{network order. The order is 2000.}
}}
\usage{
artificial_networks
}
\description{
A list called "networks" containing 200 network objects of order 2000. These
networks were simulated using the polylogarithmic (aka Gutenberg--Richter law)
degree distribution (Newman et al., 2001) with parameters
\eqn{\delta = 0.1} and \eqn{\lambda = 2} as see in the following equations:
\deqn{f(k) = k^{-{\delta}}e^{-{k/{\lambda}}}/Li_{\delta}(e^{-{1/\lambda}})}{f(k)=k^-\delta exp(-k/\lambda )/Li[\delta](exp(-1/\lambda))}
\deqn{Li=\sum_{j=1}^{\infty} z^{-j}/{j^{\delta}},}{Li=\sum_{j=1}^{\infty} z^{-j}/{j^{\delta}},}
where \eqn{\lambda > 0}. Please see reference below for details (Gel et al., 2017).
}
\references{
\insertRef{gel_bootstrap_2017}{snowboot}

\insertRef{newman_random_2001}{snowboot}
}
\keyword{datasets}
