% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/B.Bias.RMSE.ij.S.R
\name{B.Bias.RMSE.ij.S}
\alias{B.Bias.RMSE.ij.S}
\title{Estimating Bias}
\usage{
B.Bias.RMSE.ij.S(net, n.seeds, n.neigh, sam.size = 1, n.boot,
  otherNetParameters = FALSE)
}
\arguments{
\item{net}{a network object that is list containing:
\describe{
  \item{edges}{the edgelist of the network. A two column
    \code{matrix} where each row is an edge.}
  \item{degree}{the degree sequence of the network, which is
    an \code{integer} vector of length n.}
  \item{n}{the network order. The order for every network is 2000.}
}
  The object can be created by \code{\link{local.network.MR.new5}} or
  it can be imported.}

\item{n.seeds}{A vector containing the number of seeds to select during each
snowball sample. All values must be a positive integer.}

\item{n.neigh}{A vector containing the number of waves to select during each
snowball sample. All values must be a positive integer.}

\item{sam.size}{A number for the LSMI repititions. Default value is one.}

\item{n.boot}{A positive integer number, the number of bootstrap replications.}
}
\description{
A function that estimates the bias.
}

