% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Bempdegreedistrib.R
\name{Bempdegreedistrib}
\alias{Bempdegreedistrib}
\title{Obtaining an Empirical Network Degree Distribution by Bootstrap Sampling from
a Labeled Snowball Sample with Multiple Inclusion (LSMI).}
\usage{
Bempdegreedistrib(sam.out, num.sam, n.boot)
}
\arguments{
\item{sam.out}{a list that is the output of \code{\link{Oempdegreedistrib}}.}

\item{num.sam}{a vector of integers containing the number of
different samples taken from the same network.}

\item{n.boot}{\code{integer.} The size of the bootstrap sample.}
}
\value{
A list consisting of:
   \item{values}{a list of vectors containing unique
         sampled degree values. Each sample has its own vector. The output
         can be used to identify which degrees are present in the sample.}
   \item{empd}{a list of length three}
   \item{num.sam}{the number of LSMI from the \code{sam.out} object}
   \item{n.boot}{\code{integer.} The size of the bootstrap sample.}
   \item{n.neigh}{the number of waves carried out by the snowball sample.
         This is the same value from \code{sam.out$n.neigh}.}
}
\description{
The function will boostrap the a LSMI to obtain an empirical degree
distribution.
}

