% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_ci.R
\name{boot_ci}
\alias{boot_ci}
\title{Build Bootstrap Confidence Intervals from Estimators of Network Degree
Distribution.}
\usage{
boot_ci(boot_deg_out, bootstrap_mean = TRUE, alpha = 0.05)
}
\arguments{
\item{boot_deg_out}{A list that is the output of \code{\link{boot_deg}}}

\item{bootstrap_mean}{A Boolean option to return the bootstrap confidence
interval for the mean.}

\item{alpha}{\eqn{\alpha}-level of desired significance for statistical
inference. \eqn{\alpha < 0.5}.}
}
\value{
A list of two elements
 \item{p_k_CI}{list of length \code{length(boot_deg_out$num_lsmi)}, one
   element per LSMI. Each element contains bootstrap confidence
   intervals for \eqn{\hat{p}_k^*} built from the provided bootstrap
   sample of the degree distribution (supplied by the output of
   the \code{\link{boot_deg}} function).}
 \item{mean_CI}{list of length \code{length(boot_deg_out$num_lsmi)}, one
   element per LSMI. Each element contains a bootstrap confidence
   intervals for \eqn{\hat{\mu}} built from the provided bootstrap sample of
   the degree distribution (supplied by the output of
   the \code{\link{boot_deg}} function).}
}
\description{
The function will build bootstrap confidence intervals for the bootstrap
estimate of \eqn{\mu}.
}
\examples{
net <- artificial_networks[[1]]
sam.out <- empd_deg_lsmi(
  net = net,
  n.seeds = 40,
  n.neigh = 1,
  num_lsmi = 1)
boot_deg_out <- boot_deg(sam.out = sam.out, boot_rep = 50)
a <- boot_ci(boot_deg_out)
}
\references{
\insertRef{efron_bootstrap_1979}{snowboot}

\insertRef{gel_bootstrap_2017}{snowboot}
}
