% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_deg.R
\name{boot_deg}
\alias{boot_deg}
\title{Bootstrapping Empirical Degree Distribution}
\usage{
boot_deg(sam.out = sam.out, num_lsmi = sam.out$num_lsmi, boot_rep = 1)
}
\arguments{
\item{sam.out}{A list that is the output of \code{\link{empd_deg_lsmi}}.}

\item{num_lsmi}{A vector of integers containing the numeric IDs of the LSMI
samples when \code{sam.out$num_lsmi} is greater than one. When \code{num_lsmi} is an
integer, N, LSMI from 1 to N are taken from the input \code{sam.out}.}

\item{boot_rep}{A positive integer number, the number of bootstrap replications.}
}
\value{
A list consisting of:
   \item{values}{A list of length \code{num_lsmi} where each element is a
         vector containing the unique degree values sampled in each LSMI.}
   \item{empd}{A list of length \code{num_lsmi} where each element contains a
         bootstrap estimate of degree distribution for each LSMI.}
   \item{num_lsmi}{Numeric indices corresponding to LSMI samples used for bootstrap.}
   \item{boot_rep}{The same object as input argument \code{boot_rep}.}
   \item{n.neigh}{The number of waves carried out by the snowball sample.
         This is the same value from \code{sam.out$n.neigh}.}
   \item{seeds1}{A matrix of dimension \code{lenght(num_lsmi)} \eqn{\times} \code{n.seeds} with
         the numeric seed IDs. Each row corresponds to one LSMI. The rows are
         present in the same order as the IDs in \code{num_lsmi}.}
   \item{nodes_of_lsmi}{A list of length \code{length(num_lsmi)} where each
         element is vector containing the numeric IDs of the nodes sampled
         using the respective LSMI. The elements are present in the same
         order as the IDs in \code{num_lsmi}.
         Note: nodes_of_lsmi is unreported when \code{n.neigh} is zero.}
}
\description{
This function delivers a bootstrap estimate of network degree distribution
based on a LSMI sample. Default is one bootstrap replication.
}
\examples{
net <- artificial_networks[[1]]
sam.out <- empd_deg_lsmi(net = net, n.seeds = 40, n.neigh = 1, num_lsmi = 1)
a <- boot_deg(sam.out = sam.out, boot_rep = 50)
}
\references{
\insertRef{efron_bootstrap_1979}{snowboot}

\insertRef{gel_bootstrap_2017}{snowboot}
}
