% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_parameters_est.R
\name{boot_parameters_est}
\alias{boot_parameters_est}
\title{Summary of the Bootstrap Degree Distribution}
\usage{
boot_parameters_est(boot_deg_out)
}
\arguments{
\item{boot_deg_out}{A list that is the output of \code{\link{boot_deg}}}
}
\value{
A list consisting of:
   \item{mean}{An array of dimension
   \code{c(length(boot_deg_out$num_lsmi), boot_deg_out$boot_rep, 3)}.
   The last dimension, of 3, is for the three different methods of obtaining
   the empirical degree distribution from \code{boot_deg_out$empd}
   (see output empd from \code{\link{boot_deg}} for details).
   The \eqn{(i,j,k)}th element in the array is an estimate of mean degree for the
   \eqn{i}th LSMI sample, \eqn{j}th bootstrap replication, and \eqn{k}th empirical distribution
   from \code{boot_deg_out$empd}.}
   \item{quartiles}{An array of dimension
   \code{c(length(boot_deg_out$num_lsmi), 3, boot_deg_out$boot_rep, 3)}.
   The last dimension, of 3, is for the three different methods of estimation
   from \code{boot_deg_out$empd} (see output empd from \code{\link{boot_deg}}
   for details). The second dimension, of 3, corresponds to
   the quartiles (0.25, 0.5, 0.75). The \eqn{(i,j,k,l)}th element in the array is
   an estimate of \eqn{j}th quartile for the \eqn{i}th LSMI sample,
   \eqn{k}th bootstrap replication, and \eqn{l}th empirical distribution from
   \code{boot_deg_out$empd}.}
   \item{rfreq}{An array of dimension
   \code{c(length(boot_deg_out$num_lsmi), 5, boot_deg_out$boot_rep, 3)}.
   The last dimension, of 3, is for the three different methods of estimation
   from \code{boot_deg_out$empd}
   (see output empd from \code{\link{boot_deg}} for details.).
   The second dimension, of 5, corresponds to degree values: 0, 1, 2, 3, 4.
   The \eqn{(i,j,k,l)}th element in the array is the proportion of nodes
   with degree \eqn{j} in the \eqn{i}th LSMI sample, \eqn{k}th bootstrap replication,
   and \eqn{l}th empirical distribution from \code{boot_deg_out$empd}.}
   \item{deciles}{An array of dimension
         \code{c(length(boot_deg_out$num_lsmi), 9, boot_deg_out$boot_rep, 3)}.
         The last dimension, of 3, is for the three
         different methods of estimation from \code{boot_deg_out$empd}
         (see output empd from \code{\link{boot_deg}} for details.). The second
         dimension, of 9, corresponds to the deciles (0.1, 0.2, ..., 0.9).
         The \eqn{(i, j, k, l)}th element in the array is an estimate of \eqn{j}th
         decile for the \eqn{i}th LSMI sample, \eqn{k}th bootstrap replication,
         and \eqn{l}th empirical distribution from \code{boot_deg_out$empd}.}
   \item{num_lsmi}{Numeric indices corresponding to LSMI samples used for bootstrap.
         See value \code{num_lsmi} from \code{\link{boot_deg}}.}
   \item{seeds1}{A matrix of dimension \code{length(num_lsmi)} \eqn{\times} \code{n.seeds} with
         the numeric seed IDs. Each row corresponds to one LSMI. The rows are
         present in the same order as the IDs in \code{num_lsmi}.
         See value \code{seeds1} from \code{\link{boot_deg}}.}
   \item{nodes_of_lsmi}{A list of length \code{length(num_lsmi)} where each
         element is vector containing the numeric IDs of the nodes sampled
         using the respective LSMI. The elements are present in the same
         order as the IDs in \code{num_lsmi}.
         Note: \code{nodes_of_lsmi} is unreported when n.neigh equals zero.
         See value \code{nodes_of_lsmi} from \code{\link{boot_deg}}.}
}
\description{
This function provides summary statistics of a bootstrap degree
distribution.
}
\examples{
net <- artificial_networks[[1]]
sam.out <- empd_deg_lsmi(net = net, n.seeds = 40, n.neigh = 1, num_lsmi = 1)
boot_deg_out <- boot_deg(sam.out = sam.out, boot_rep = 50)
a <- boot_parameters_est(boot_deg_out)
}
\references{
\insertRef{efron_bootstrap_1979}{snowboot}

\insertRef{gel_bootstrap_2017}{snowboot}
}
