% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCI.R
\name{bootCI}
\alias{bootCI}
\title{Build Bootstrap Confidence Intervals for \eqn{\hat{p}_k^*}}
\usage{
bootCI(outBootdeg, bootstrap_mean = T)
}
\arguments{
\item{outBootdeg}{a list that is the output of \code{\link{bootdeg}}}
}
\value{
A list of two elements
}
\description{
The function will build bootstrap confidence intervals for the bootstrap
estimate of  and \eqn{\mu} with a lower-bound of
\code{0.025} and an upper-bound of \code{0.975}.
}

