% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation_mean.R
\name{cross_validation_mean}
\alias{cross_validation_mean}
\title{A function that uses cross-validation to select seed-wave combination of LSMI
for the estimated mean degree.}
\usage{
cross_validation_mean(network, n.seeds, n.neigh, boot_rep, alpha = 0.05,
  proxyRep = 19, proxyOrder = 30)
}
\arguments{
\item{network}{A network object that is list containing:
\describe{
  \item{edges}{The edgelist of the network. A two column
    \code{matrix} where each row is an edge.}
  \item{degree}{The degree sequence of the network, which is
    an \code{integer} vector of length n.}
  \item{n}{The network order.}
}
  The object can be created by \code{\link{random_network}} or
  it can be imported.}

\item{n.seeds}{A numeric vector for the different sample sizes of seed to use
in cross-validation.}

\item{n.neigh}{A numeric vector for the different waves to use
in cross-validation.}

\item{boot_rep}{The number of bootstrap sample.}

\item{alpha}{Desired type I error for bootstrap confidence intervals, which
are obtained using the quantile method.}

\item{proxyRep}{The number of time to sample a proxy. Default is 19.}

\item{proxyOrder}{The size of the proxy sample. Default is 30.}
}
\value{
A list consisting of
 \item{best_combination}{A matrices that provides
   the best seed-wave combinations for LSMI (obtained via cross-validation)
   when estimating mean degree.}
 \item{best_boot_ci_mean}{A vector of length 2 that provides
   the bootstrap confidence intervals for the estimated mean degree
   using the best seed-wave combinations (see above).}
}
\description{
The function's inputs are a network, a vector of possible seed sample-sizes,
a vector of possible waves, and a few tuning parameters. The output will
contain the best seed-wave combination for the mean degree and the width of the
bootstrap confidence intervals for the mean degree when using
the best seed-wave combination.
}
\note{
Only one LSMI per seed-wave combination is currently supported.
}
\examples{
net <- artificial_networks[[1]]
a <- cross_validation_mean(network = net, n.seeds = c(10, 20, 30), n.neigh = c(1, 2),
 boot_rep = 200)
}
\references{
\insertRef{efron_bootstrap_1979}{snowboot}

\insertRef{gel_bootstrap_2017}{snowboot}
}
