% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraph_to_network.R
\name{igraph_to_network}
\alias{igraph_to_network}
\title{Bootstraping Empirical Degree Distribution}
\usage{
igraph_to_network(in_graph)
}
\arguments{
\item{in_graph}{an igraph object. To create igraph objects from field data,
see \code{\link[igraph]{graph_from_edgelist}},
\code{\link[igraph]{graph_from_data_frame}},
\code{\link[igraph]{graph_from_adjacency_matrix}}, or
\code{\link[igraph]{read_graph}}.}
}
\value{
a list that contain elements:
   \item{edges}{the edgelist of the network. A two column
     \code{matrix} where each row is an edge.}
   \item{degree}{the degree sequence of the network, which is
     an \code{integer} vector of length n.}
   \item{n}{the network order.}
}
\description{
This function delivers a boostrap estimate of network degree distribution
based on a LSMI sample. Default is one bootstrap replication.
}
\references{
\url{http://igraph.org/}
}

