% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampleneighAroundOneSeed.R
\name{sampleneighAroundOneSeed}
\alias{sampleneighAroundOneSeed}
\title{A function to snowball sample the neighbors around a single seed.}
\usage{
sampleneighAroundOneSeed(net, seed0, n.neigh = 1)
}
\arguments{
\item{net}{A network object.}

\item{seed0}{A number that is the id of the seed whose neighbors are sampled.}

\item{n.neigh}{A number for the waves of snowball sampling.}
}
\value{
a list containing:
   \item{seed}{A number that is the id of the seed whose neighbors are sampled.}
   \item{sampleN}{A numeric vector containg ids of the vertices from
         the snowball sampling and the intial seed's id. This vector may have
         duplicates, since the algorithm allows for multiple inclusions.}
   \item{unode}{A numeric vector containg the unique values in \code{$sampleN}.}
   \item{nodes.waves}{A list of numeric vectors (one per wave) each
         containing the ids of the nodes sampled in that particular wave.}
}
\description{
This function returns the vertices samples by snowball up to wave n.neigh
around a single seed. The functions main purpose is to be called from
\code{\link{sampleneighSequential}.}
}

