% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampleneighSequential.R
\name{sampleneighSequential}
\alias{sampleneighSequential}
\title{A function for snowball sampling.}
\usage{
sampleneighSequential(net, n.seeds = 10, n.neigh = 1, seed = NULL)
}
\arguments{
\item{net}{A network object.}

\item{n.seeds}{A number designating the desired number of seeds.}

\item{n.neigh}{A number designating teh desired number of waves.}

\item{An}{optional \code{numeric} vector with desired seed ids.}
}
\value{
A list containing:
   \item{seeds}{A \code{numeric} vector containing the ids of the seeds.}
   \item{sampleN}{A numeric vector containg ids of the vertices from
         the snowball sampling and the intial seeds' ids. This vector may have
         duplicates, since the algorithm allows for multiple inclusions.}
   \item{unodes}{a list of \code{numeric} vectors (one per seed) each
         containing the seed's id and the unique ids of all nodes that were
         snowball sampled from that seed using
         \code{\link{sampleneighAroundOneSeed}}.}
   \item{nodes.waves}{A list of list (one per seed) each containing, yet
         another, list of numeric vectors (one per wave) each
         containing the ids of the nodes sampled in that particular wave. The
         second outter most list is the output element \code{$nodes.waves}
         from \code{\link{sampleneighAroundOneSeed}}.}
}
\description{
The function will conduct snowball sampling on a network object when given
the network object, number of desired seeds (or the actual seed ids), and
the desired number of waves.
}

