% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_k_cross_validation.R
\name{sort_tied_opti}
\alias{sort_tied_opti}
\title{A function that sorts a matrix of tied optimal seed-wave combinations so that
the first is the largest number of seeds among the smallest waves.}
\usage{
sort_tied_opti(inMat)
}
\arguments{
\item{inMat}{is a matrix of tied seed-wave indices.}
}
\value{
A matrix with rows such that the indices for the largest number of
seeds (first column) for the lowest number of waves (second column) is on top.
}
\description{
The function takes a matrix containing tied optimal seed-wave combinations
from the training proxy and sorts the rows of this matrix so that the
largest number of seeds for the lowest number of waves is on top
}

