% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.net.R
\name{summary.net}
\alias{summary.net}
\title{Summary of a Network Degree Sequence}
\usage{
summary.net(net)
}
\arguments{
\item{net}{a network object that is list containing:
\describe{
  \item{edges}{the edgelist of the network. A two column
    \code{matrix} where each row is an edge.}
  \item{degree}{the degree sequence of the network, which is
    an \code{integer} vector of length n.}
  \item{n}{the network order. The order for every network is 2000.}
}
  The object can be created by \code{\link{local.network.MR.new5}} or
  it can be imported.}
}
\value{
a list consisting of:
   \item{realdd}{a vector of length \code{net$n} where each element
   corresponds to the degree of a node in the network.}
   \item{rmean}{a numeric vector of length one that is the atrithmetic mean
         of \code{realdd}. see (\code{\link[base]{mean}}).}
   \item{rquart}{the lower, median, and upper quartiles of degree sequence
         \code{realdd}.}
   \item{rfreq}{a numeric vector of length five where each element
         corresponds to the proportion of seeds with degree: 0, 1, 2, 3, and
         4 (respectively).}
   \item{rdeci}{a numeric vector of lenth nine containing the deciles of
         of degree sequence \code{realdd} in increasing order.}
}
\description{
This function provides summary statistics of a network degree
distribution.
}

