% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCI.R
\name{bootCI}
\alias{bootCI}
\title{Build Bootstrap Confidence Intervals for \eqn{\hat{p}_k^*}}
\usage{
bootCI(outBootdeg, bootstrap_mean = T)
}
\arguments{
\item{outBootdeg}{a list that is the output of \code{\link{bootdeg}}}

\item{bootstrap_mean}{a boolean option to return the bootstrap confidence
interval for the mean.}
}
\value{
A list of two elements
 \item{p_k_CI}{This a list of length \code{length(outBootdeg$num.sam)}, one
   element per LSMI. Each element is contains three sets of bootstrap confidence
   intervals for \eqn{\hat{p}_k^*} corresponding to the three estimation
   methods. See \code{\link{bootdeg}} for more on the three estimation methods.}
 \item{mean_CI}{This a list of length \code{length(outBootdeg$num.sam)}, one
   element per LSMI. Each element is contains three sets of bootstrap confidence
   intervals for \eqn{\hat{\mu}} corresponding to the three estimation
   methods. See \code{\link{bootdeg}} for more on the three estimation methods.}
}
\description{
The function will build bootstrap confidence intervals for the bootstrap
estimate of  and \eqn{\mu} with a lower-bound of
\code{0.025} and an upper-bound of \code{0.975}.
}

