% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootCI.R
\name{bootCI}
\alias{bootCI}
\title{Build Bootstrap Confidence Intervals for \eqn{\hat{p_{k}}}

The function will build bootstrap confidence intervals for the bootstrap
estimate of \eqn{\mu}.}
\usage{
bootCI(outBootdeg, bootstrap_mean = TRUE, alpha = 0.05)
}
\arguments{
\item{outBootdeg}{A list that is the output of \code{\link{bootdeg}}}

\item{bootstrap_mean}{A Boolean option to return the bootstrap confidence
interval for the mean.}

\item{alpha}{\eqn{\alpha}-level of desired significance for statistical
inference. \eqn{\alpha < 0.5}.}
}
\value{
A list of two elements
 \item{p_k_CI}{list of length \code{length(outBootdeg$num.sam)}, one
   element per LSMI. Each element contains bootstrap confidence
   intervals for \eqn{\hat{p}_k^*} built from the provided bootstrap  sample of
   the degree distribution (supplied by the output of
   the \code{\link{bootdeg}} function). See \code{\link{bootdeg}} for more
   on the available estimation methods.}
 \item{mean_CI}{list of length \code{length(outBootdeg$num.sam)}, one
   element per LSMI. Each element contains a bootstrap confidence
   intervals for \eqn{\hat{\mu}} built from the provided bootstrap sample of
   the degree distribution (supplied by the output of
   the \code{\link{bootdeg}} function). See \code{\link{bootdeg}} for more
   on the available estimation methods.}
}
\examples{
net <- artificial_networks[[1]]
sam.out <- Oempdegreedistrib(net = net, n.seeds = 40, n.neigh = 1, num.sam = 1)
outBootdeg <- bootdeg(sam.out = sam.out, n.boot = 50)
a <- bootCI(outBootdeg)
}
\references{
Efron, B. (1992). Bootstrap methods: another look at the jackknife.
In Breakthroughs in statistics (pp. 569-593). Springer New York.

Gel, Y. R., Lyubchich, V., & Ramirez Ramirez, L. L. (2017).
Bootstrap quantification of estimation uncertainties in
network degree distributions. Scientific Reports, 7, 5807.
\url{http://doi.org/10.1038/s41598-017-05885-x}
}
