% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bootdeg.R
\name{bootdeg}
\alias{bootdeg}
\title{Bootstraping Empirical Degree Distribution}
\usage{
bootdeg(sam.out, num.sam = sam.out$num.sam, n.boot = 1)
}
\arguments{
\item{sam.out}{a list that is the output of \code{\link{Oempdegreedistrib}}.}

\item{num.sam}{a vector of integers containing the numeric ids of the LSMI
samples when \code{sam.out$num.sam} is greater than one. When it is an
integer, N, , all LSMI from 1 to N are taken from the input \code{sam.out}.}

\item{n.boot}{A positive integer number, the number of bootstrap replications.}
}
\value{
A list consisting of:
   \item{values}{a list of length \code{num.sam} where each element is a
         vector containing the unique degree values sampled in each LSMI.}
   \item{empd}{a list of length three where each element contains a different
         estimate of degree distribution: empd.w.p0s - weighted bootstrap
         with a proportion of isolated nodes p0 being estimated by simple
         random sampling of bootstrapped seeds; empd.nw.p0sEkb - non-weighted
         bootstrap with a proportion of isolated nodes p0 being estimated by
         simple random sampling of bootstrapped seeds; empd.nw.p0sEks -ignore
         (see Thompson et al. for details)}
   \item{num.sam}{the same object as input argument \code{num.sam}}
   \item{n.boot}{the same object as input argument \code{n.boot}.}
   \item{n.neigh}{the number of waves carried out by the snowball sample.
         This is the same value from \code{sam.out$n.neigh}.}
   \item{seeds1}{a matrix of dimension \code{num.sam} x \code{n.seeds} with
         the numeric seed ids. Each row corresponds to one LSMI. However,
         when \code{num.sam} is a vector of the desired LSMI, the matrix has
         dimension \code{length(num.sam)} x \code{n.seeds}. In this case,
         each row corresponds to the LSMIs denoted in the vector
         \code{num.sam}, and the rows are present in the same order as the
         ids in \code{num.sam}.}
   \item{nodes_of_LSMI}{a list of length \code{num.sam} where each element is
         vector containing the numeric ids of the nodes sampled using LSMI
         (One element per LSMI). However, when \code{num.sam} is a vector of
         the desired LSMI, the list has length \code{length(num.sam)}.
         In this case, each element corresponds to the LSMIs denoted in
         the vector \code{num.sam}, and the elements are present in the same
         order as the ids in \code{num.sam}.
         Note: nodes_of_LSMI is unreported when n.neigh equals zero.}
}
\description{
This function delivers a boostrap estimate of network degree distribution
based on a LSMI sample. Default is one bootstrap replication.
}

