% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BparametersEst.R
\name{BparametersEst}
\alias{BparametersEst}
\title{Summary of a Network Degree Sequence}
\usage{
BparametersEst(outBootdeg)
}
\arguments{
\item{outBootdeg}{a list that is the output of \code{\link{bootdeg}}}
}
\value{
a list consisting of:
   \item{mean}{an array of dimension \code{length(outBootdeg$num.sam) x
         outBootdeg$n.boot x 3}. The last dimension, of 3, is for the three
         different methods of obtaining the empirical degree distribution
         from \code{outBootdeg$empd}
         (see output empd from \code{\link{bootdeg}} for details.).
         Each element in the array is an estimation of the mean for the
         respective LSMI, bootstrap replication, and empirical distribution
         from \code{outBootdeg$empd}.}
   \item{quartiles}{an array of dimension \code{length(outBootdeg$num.sam) x 3 x
         outBootdeg$n.boot x 3}. The last dimension, of 3, is for the three
         different methods of estimation from \code{outBootdeg$empd}
         (see output empd from \code{\link{bootdeg}} for details.). The second
         dimension, of 3, corresponds to the quartiles (.25, .5, .75).
         Each element in the array is a quartile for the
         respective LSMI, quartile (.25, .5, .75), for a bootstrap replication,
         and empirical distribution from \code{outBootdeg$empd}.}
   \item{rfreq}{an array of dimension \code{length(outBootdeg$num.sam) x 5 x
         outBootdeg$n.boot x 3}. The last dimension, of 3, is for the three
         different methods of estimation from \code{outBootdeg$empd}
         (see output empd from \code{\link{bootdeg}} for details.). The second
         dimension, of 5, corresponds to degree values: 0, 1, 2, 3, 4.
         Each element in the array is the proportion of nodes from the
         respective LSMI, with degree value (0-4), for a bootstrap replication,
         and empirical distribution from \code{outBootdeg$empd}.}
   \item{deciles}{an array of dimension \code{length(outBootdeg$num.sam) x 9 x
         outBootdeg$n.boot x 3}. The last dimension, of 3, is for the three
         different methods of estimation from \code{outBootdeg$empd}
         (see output empd from \code{\link{bootdeg}} for details.). The second
         dimension, of 9, corresponds to the deciles (.1, .2, ... , .9).
         Each element in the array is a decile for the
         respective LSMI, decile (.1, .2, ... , .9), for a bootstrap replication,
         and empirical distribution from \code{outBootdeg$empd}.}
   \item{num.sam}{same as \code{outBootdeg$num.sam}}
}
\description{
This function provides summary statistics of a network degree
distribution.
}

