% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sampleneighSequential.R
\name{sampleneighSequential}
\alias{sampleneighSequential}
\title{Snowball sampling with multiple inclusion.}
\usage{
sampleneighSequential(net, n.seeds = 10, n.neigh = 1, seeds = NULL)
}
\arguments{
\item{net}{a list that must contain elements
\code{$n} (\code{integer}. network order),
\code{$edges} (\code{matrix}. a \code{n}x2 matrix),
and \code{$degree} (\code{integer} of length n).
The object can be created by \code{\link{local.network.MR.new5}} or
it can be imported.}

\item{n.seeds}{a number of seeds in the snowball sample.
It must be a positive integer.}

\item{n.neigh}{a number of waves to be sampled around each seed in LSMI.
For example, n.neigh = 0 corresponds to seeds only, and n.neigh = 1
corresponds to sampling seeds and their first neighbors).
Note that the algorithm allow for mutiple inclusions.}

\item{seeds}{an optional paramater that is used whenever we know the seeds
   we would like to initiate the snowball with. WARNING: If this parameter
   is used, then num.sam > 1, this will
   cause an error.}
}
\value{
A list containing the following elements:
   \item{seeds}{A \code{numeric} a vector containing the numerics ids of
         sampled seeds.}
   \item{sampleN}{A numeric vector containg ids of the nodes from
         the snowball sampling and the intial seeds' ids. This vector may have
         duplicates, since the algorithm allows for multiple inclusions.}
   \item{unodes}{a list of \code{numeric} vectors (one per seed) each
         containing the seed's id and the unique ids of all nodes that were
         snowball sampled from that seed using
         \code{\link{sampleneighAroundOneSeed}}.}
   \item{nodes.waves}{A list of list (one per seed) each containing, yet
         another, list of numeric vectors (one per wave) each
         containing the ids of the nodes sampled in that particular wave. The
         second outter most list is the output element \code{$nodes.waves}
         from \code{\link{sampleneighAroundOneSeed}}.}
}
\description{
The function will conduct snowball sampling.
}

